feat: depreciate cartesi build command
