+++ apps/cli/src/commands/rollups.ts
@@ -1,4 +1,5 @@
 import { Command } from "@commander-js/extra-typings";
+import { createBuildCommand } from "./rollups/build.js";
 import { createCreateCommand } from "./rollups/create.js";
 import { createDeployCommand } from "./rollups/deploy.js";
 import { createLogsCommand } from "./rollups/logs.js";
@@ -22,6 +23,7 @@ export const createRollupsCommand = () => {
     command.addCommand(createStopCommand());
     command.addCommand(createDeployCommand());
     command.addCommand(createCreateCommand());
+    command.addCommand(createBuildCommand());
     return command;
 };
